Only import external JS and CSS once
With JS Booster support added, templates that imported external JS and CSS became an issue as they would get imported again and again with each added card preview. As a result, performance would diminish very quickly. This commit makes sure that external scripts and stylesheets are only imported once.

One minor issue: This assumes that all cards depend on the same external files, which doesn't have to be the case. Ideally the add-on would only import each unique script or stylesheet once, but this would require a much more complex regex match and logic. Marking this as TODO.