Move load_config() to page rendering stage; refactoring
'profileLoaded' is only run after drawing the deck browser for
the first time which makes it unsuitable for setting up the
initial add-on configuration. Having load_config() run any time
we move to the deckbrowser or overview stage is less than ideal,
but it seems to be the only viable solution for now.